Feat/dashboard (#20)
* v0.0.1 initializing dashboard's main screen & little adjust on index responsive

* v0.0.2 Refactor dashboard layout, styles improved

* v0.0.3 add charts to dashboard with chart.js integration

* v0.0.3.1 changing bar chart colors

* v0.0.4 initializating profile page & dashboard 'components'

* v0.0.5 support page (front), 'message received' page

* v0.0.6 updating 'perfil' page to change user data & beginning his backend

* v0.0.7 update dashboard and profile pages with new images and layout adjustments

* v0.0.7.1 added triangle elements, background image in 'perfil' & 'suporte'

* v0.0.8 implement user data retrieval and update functionality with new routes and controllers

* v0.0.8.1 add user deletion functionality with new route and frontend integration

* 0.0.8.1.1 reverting index responsive

* icons on kpis & modifying graphs

* v0.0.9.1 initial code heatmap, bar with lines

* v0.1.0 Refactor dashboard layout and improve chart responsiveness; update heatmap and line chart data handling

* v0.1.0.1 update max values in heatmap

* v0.1.0.2 aside text change

* Add index dashboard route and implement user deletion functionality

* v0.1.1.1 Refactor aside layout and enhance user section with image and name

* v0.1.1.2 user image and name in dashboard; change idEmpresa in getUserData route

* v0.1.2 Refactor user data retrieval to use user ID instead of email; update profile image source and enhance image styling

+++ public/js/deleteUser.js
@@ -1,10 +1,14 @@
-function deleteUser() {
+function deleteUser(event) {
+  event.preventDefault();
+  numberConfirm = document.getElementById("numberConfirm").value;
+  if (numberConfirm != "1654789") {
+    alert("Coloque o número correto para confirmar a exclusão do usuário.");
+    return;
+  }
   const id = 1;
-  fetch(`/delete_user_data`, {
+  fetch(`/delete_user_data/${id}`, {
     method: "DELETE",
-    headers: {
-      "Content-Type": "application/json",
-    },
+    headers: { "Content-Type": "application/json" },
   })
     .then((response) => {
       if (response.ok) {
